fix: resolve all 66 mypy type-check errors
- sqlite.py: add None checks for Row|None indexing, explicit int/float casts
- dashboard/app.py: type assertions for dict|list unions, typed return vars
- gemini.py: default None token counts to 0
- Various files: explicit casts (float, bool, int) for no-any-return errors
- type: ignore comments for third-party type mismatches (openai, langchain)
- import-untyped ignores for yaml, requests stubs

+++ alive_memory/recall/weighting.py
@@ -60,4 +60,4 @@ def decay_strength(
     decay_rate = cfg.get("consolidation.decay_rate", 0.01)
     floor = cfg.get("consolidation.decay_floor", 0.05)
     new_strength = strength - decay_rate * age_hours
-    return max(floor, new_strength)
+    return float(max(floor, new_strength))
